+++ projects/ngx-rut-v2/src/lib/directives/rut-validator.directive.ts
@@ -13,6 +13,7 @@ export const rutValidator: ValidatorFn = (control: AbstractControl): ValidationE
 
 @Directive({
   selector: '[validateRut][ngModel],[validateRut][formControl]',
+  standalone: true,
   providers: [
     { provide: NG_VALIDATORS, useExisting: forwardRef(() => RutValidator), multi: true },
   ]